Update balena.yml
Hey @Razikus, we've recently updated the syntax for the balena.yml file and the concept of balenaHub apps (fleets and projects). Details can be found [here](https://www.balena.io/blog/introducing-open-fleets-and-self-submitted-apps-and-blocks-on-balenahub/).

I made the necessary changes in this PR, feel free to change what you think is necessary. Let's stay connected!

Change-type: patch
Signed-off-by: Marc Pous marc@balena.io